requirements.txt
tweepy==3.6.0
discord.py==0.16.12
discord==1.0.1
PyYAML==5.1.1
